requirements.txt
Django>=3.0.0,<3.1.0
djangorestframework>=3.12.4,<3.13.0
